Resolve Expression
Resolving issue that when passing a named value (i.e. gamma or nu) to expression, it cannot evaluate it when differentiating the expression. Trying to include them as variables to the function so that they can be evaluated for the differentiation without being treated as parameters. This introduces new parameters to the sandwhich estimator, and new expressions in Pareto Analysis