HSandwich
Expression terms were not working while trying to pass values of gamma to the expression for the evaluation of the score function or the bordered Hessian function. Therefore, created a new version of the sandwich estimated called HSandwich (or HardSandwich) that hard codes the equations into the sandwich estimator to make sure that the calculation can be made.